quote: Sometimes we try to escape from some technology, but it's no use!
It's common at the beginning of a software engineering career to want to *"escape"* from certain technology stacks (for example JavaScript, Java, etc).

After I understood that **adding value is the delivery to the business**, it made me not care about the stack I'm using.

I enjoy programming and it's possible to write good software **regardless of the stack**.

Signed-off-by: Avelino <31996+avelino@users.noreply.github.com>